Rel: Inicia implementação de relatório

+++ src/main/java/com/loiane/model/Aluno.java
@@ -25,7 +25,7 @@ public class Aluno {
     private String cpf;
 
     @Column(scale=2)
-    private Double idade;
+    private Integer idade;
 
     @NotNull
     @Length(max = 100)
